printf支持输出: printf(str,type)

+++ src/utils.cpp
@@ -1,10 +1,10 @@
 #include <string>
 
 int str_to_int(const std::string& s) {
-	return std::atoi(s.c_str());
+	return std::stoi(s.c_str());
 }
 float str_to_float(const std::string& s) {
-	return float(std::atof(s.c_str()));
+	return std::stof(s.c_str());
 }
 void replace_str(std::string& str, std::string oldstr, std::string newstr) {
 	size_t cur_pos	= 0;
